[SPARK-46471][PS][TESTS][FOLLOWUPS] Reorganize `OpsOnDiffFramesEnabledTests`: Factor out more tests
### What changes were proposed in this pull request?
factor out following tests:

- test_bitwise
- test_bitwise_extension_dtype
- test_combine_first
- test_compare
- test_concat_column_axis -> test_concat_column_axis_inner & test_concat_column_axis_outer

### Why are the changes needed?
for testing parallelism

### Does this PR introduce _any_ user-facing change?
no

### How was this patch tested?
ci

### Was this patch authored or co-authored using generative AI tooling?
no

Closes #44469 from zhengruifeng/ps_test_diff_ops_2.

Authored-by: Ruifeng Zheng <ruifengz@apache.org>
