+++ dev/sparktestsupport/modules.py
@@ -872,6 +872,11 @@ pyspark_pandas_slow = Module(
         "pyspark.pandas.tests.diff_frames_ops.test_arithmetic_chain_ext_float",
         "pyspark.pandas.tests.diff_frames_ops.test_assign_frame",
         "pyspark.pandas.tests.diff_frames_ops.test_assign_series",
+        "pyspark.pandas.tests.diff_frames_ops.test_bitwise",
+        "pyspark.pandas.tests.diff_frames_ops.test_combine_first",
+        "pyspark.pandas.tests.diff_frames_ops.test_compare_series",
+        "pyspark.pandas.tests.diff_frames_ops.test_concat_inner",
+        "pyspark.pandas.tests.diff_frames_ops.test_concat_outer",
         "pyspark.pandas.tests.diff_frames_ops.test_basic_slow",
         "pyspark.pandas.tests.diff_frames_ops.test_cov",
         "pyspark.pandas.tests.diff_frames_ops.test_corrwith",
@@ -1239,6 +1244,11 @@ pyspark_pandas_connect_part3 = Module(
         "pyspark.pandas.tests.connect.diff_frames_ops.test_parity_arithmetic_chain_ext_float",
         "pyspark.pandas.tests.connect.diff_frames_ops.test_parity_assign_frame",
         "pyspark.pandas.tests.connect.diff_frames_ops.test_parity_assign_series",
+        "pyspark.pandas.tests.connect.diff_frames_ops.test_parity_bitwise",
+        "pyspark.pandas.tests.connect.diff_frames_ops.test_parity_combine_first",
+        "pyspark.pandas.tests.connect.diff_frames_ops.test_parity_compare_series",
+        "pyspark.pandas.tests.connect.diff_frames_ops.test_parity_concat_inner",
+        "pyspark.pandas.tests.connect.diff_frames_ops.test_parity_concat_outer",
         "pyspark.pandas.tests.connect.diff_frames_ops.test_parity_groupby",
         "pyspark.pandas.tests.connect.diff_frames_ops.test_parity_groupby_aggregate",
         "pyspark.pandas.tests.connect.diff_frames_ops.test_parity_groupby_apply",
